abstract_uart: Refactor API for poll-based termination
