+++ vp/src/platform/common/uart.cpp
@@ -8,17 +8,10 @@
 
 #include <sys/types.h>
 
-#define stop_fd (stop_pipe[0])
-#define newpollfd(FD) \
-	(struct pollfd){.fd = FD, .events = POLLIN | POLLERR};
-
 UART::UART(const sc_core::sc_module_name& name, uint32_t irqsrc)
 		: AbstractUART(name, irqsrc) {
-	fds[0] = newpollfd(stop_fd);
-	fds[1] = newpollfd(STDIN_FILENO);
-
 	enableRawMode(STDIN_FILENO);
-	start_threads();
+	start_threads(STDIN_FILENO);
 }
 
 UART::~UART(void) {
@@ -36,7 +29,6 @@ void UART::handle_input(int fd) {
 		throw std::runtime_error("short read");
 
 	rxpush(buf);
-
 }
 
 void UART::write_data(uint8_t data) {
@@ -47,22 +39,4 @@ void UART::write_data(uint8_t data) {
 		throw std::system_error(errno, std::generic_category());
 	else if (nwritten != sizeof(data))
 		throw std::runtime_error("short write");
-}
-
-void UART::read_data(void) {
-	if (poll(fds, (nfds_t)NFDS, -1) == -1)
-		throw std::system_error(errno, std::generic_category());
-
-	/* stop_fd is checked first as it is fds[0] */
-	for (size_t i = 0; i < NFDS; i++) {
-		int fd = fds[i].fd;
-		short ev = fds[i].revents;
-
-		if (fd == stop_fd)
-			break;
-		else if (ev & POLLERR)
-			throw std::runtime_error("received unexpected POLLERR");
-		else
-			handle_input(fd);
-	}
 }
