+++ vp/src/platform/common/uart.h
@@ -2,7 +2,6 @@
 #define RISCV_VP_UART_H
 
 #include <stdint.h>
-#include <poll.h>
 #include <systemc>
 #include "abstract_uart.h"
 
@@ -12,14 +11,8 @@ public:
 	~UART(void);
 
 private:
-	enum {
-		NFDS = 2,
-	};
-	struct pollfd fds[NFDS];
-
-	void handle_input(int fd);
-	void write_data(uint8_t);
-	void read_data(void);
+	void handle_input(int fd) override;
+	void write_data(uint8_t) override;
 };
 
 #endif  // RISCV_VP_UART_H
